Move StructuralReference to new interfaces file
