+++ fmriprep/interfaces/__init__.py
@@ -4,4 +4,5 @@
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 from fmriprep.interfaces.bids import ReadSidecarJSON, DerivativesDataSink, \
     BIDSDataGrabber, BIDSFreeSurferDir
-from fmriprep.interfaces.images import IntraModalMerge, CopyHeader, StructuralReference
+from fmriprep.interfaces.images import IntraModalMerge, CopyHeader
+from fmriprep.interfaces.freesurfer import StructuralReference
